Implement MaxFlowSolver for image denoising
- Add the `MaxFlowSolver` class carry out image denoising
  using Boykov-Kolmogorov Max-Flow algorithm.
- Add the `MaxFlowSolverImplementation` class
  that uses The Boost Graph library.

Find details about the Boykov-Kolmogorov Max-Flow implementation
in The Boost Graph Library here:
https://www.boost.org/doc/libs/1_80_0/libs/graph/doc/boykov_kolmogorov_max_flow.html

The implementation uses CSR because it suits the task well.
Its initialisation is performed significantly faster
than for Adjacency List,
and it is far more memory efficient than
an ordinary Adjacency Matrix.
For more information, read
https://www.boost.org/doc/libs/1_80_0/libs/graph/doc/compressed_sparse_row.html

The implementation class uses private functions
to construct the graph because it needs to be preallocated.
As long as we know the image size beforehand,
we can do this.

Use custom exceptions to inspect the application failures better.