Add GitHub Action for automated builds
Add an action that installs dependencies
and compiles the project with `g++` and `clang++`.
Check both Debug and Release modes.

Currently, there are no tests, so `ctest -C` action is omitted.

Install `libboost-graph-dev`
because the project needs only The Graph Boost Library,
not the entire Boost.

Install `libopencv-dev`
because it contains the files
necessary for CMake project (Ubuntu 22, Jammy):
https://packages.ubuntu.com/jammy/amd64/libopencv-dev/filelist